+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.6
+* Fix ExAlert, ExSnackbar, ExLoading
+
+
+## 0.0.4
+* remove parameter context from features (ex_alert, ex_snackbar, etc)
+
 ## 0.0.3
 * add ext isNotNullOrEmpty